+++ Tests/FluentAssertions.Specs/Primitives/BooleanAssertionSpecs.cs
@@ -41,7 +41,7 @@ public class BooleanAssertionSpecs
             // Assert
             action
                 .Should().Throw<XunitException>()
-                .WithMessage("Expected boolean to be true because we want to test the failure message, but found False.");
+                .WithMessage("Expected boolean to be True because we want to test the failure message, but found False.");
         }
     }
 
@@ -78,7 +78,7 @@ public class BooleanAssertionSpecs
 
             // Assert
             action.Should().Throw<XunitException>()
-                .WithMessage("Expected boolean to be false because we want to test the failure message, but found True.");
+                .WithMessage("Expected boolean to be False because we want to test the failure message, but found True.");
         }
     }